refactor: create a separate component for word typing
- move it to its own folder, and update imports
