+++ src/app/home/home.component.html
@@ -1,34 +1,3 @@
 <div class="container">
-	<!--TODO: create a seperate component for this-->
-	<div class="word-box" (click)="userInput.focus()">
-		<div class="prev-words">
-			<div
-				class="prev-word"
-				*ngFor="let prevWord of previousWords; index as i"
-			>
-				<span *ngFor="let char of prevWord.split(''); index as c">{{
-					char
-				}}</span>
-			</div>
-		</div>
-		<span class="blinking-cursor">|</span>
-		<div class="curr-words">
-			<div
-				class="curr-word"
-				*ngFor="let word of currentWords; index as i"
-			>
-				<span *ngFor="let char of word.split(''); index as c">{{
-					char
-				}}</span>
-			</div>
-		</div>
-	</div>
-	<input
-		type="text"
-		class="user-input"
-		#userInput
-		(keypress)="onKeyPress($event)"
-		(keydown.space)="onSpaceKeyPress(); userInput.value = ''"
-		(keydown.backspace)="onBackspaceKeyPress()"
-	/>
+	<app-word-typing class="words"></app-word-typing>
 </div>
