+++ src/app/home/word/word-typing.component.html
@@ -1,0 +1,31 @@
+<div class="word-box" (click)="userInput.focus()">
+		<div class="prev-words">
+			<div
+				class="prev-word"
+				*ngFor="let prevWord of previousWords; index as i"
+			>
+				<span *ngFor="let char of prevWord.split(''); index as c">{{
+					char
+				}}</span>
+			</div>
+		</div>
+		<span class="blinking-cursor">|</span>
+		<div class="curr-words">
+			<div
+				class="curr-word"
+				*ngFor="let word of currentWords; index as i"
+			>
+				<span *ngFor="let char of word.split(''); index as c">{{
+					char
+				}}</span>
+			</div>
+		</div>
+	</div>
+	<input
+		type="text"
+		class="user-input"
+		#userInput
+		(keypress)="onKeyPress($event)"
+		(keydown.space)="onSpaceKeyPress(); userInput.value = ''"
+		(keydown.backspace)="onBackspaceKeyPress()"
+/>
